Some changes to how SQL is handled.
Table structure is as follows:

- String world
- long location
- String sprout
- int age

Plugin will alert admin if Sprouts are being attempted to be dispersed
into the world without equilvalent base objects in the registry. The
plugin also drops the entire memory storage to SQL at shutdown now. Should
be interesting to see the performance results of this.

Signed-off-by: Chris Sanders <zidane@outlook.com>