Fixed material lookup issues, added placement.
You can now add a placement-source under the item-source for your sprout
in sprouts.yml to specify what block the seed must have to be planted in
the world.

This allows cool things such as having a crop that only grows in
cobblestone. Bear in mind, we cannot stop the Minecraft Server from
uprooting the base block so choose a compatible base block within
MoreMaterials (or whereever your custom item is coming from).

We find Rose to be the most compatibile. Lastly, this is an OPTIONAL
change, by default sprouts need soil.
Signed-off-by: Chris Sanders <zidane@outlook.com>